Run all tests when the enter key is pressed
A developer should be able to trigger a full test run arbitrarily under continuous testing mode.
This applies mostly to situations where tests where ran for a subset of the target namespaces
that where subject to change and we want to check that these changes didn't brake something outside
of the code covered by the restricted run.